Move IOError exception handling
IOErrors were not being caught in the right place.  They happen in the read
loop, not the send loop.  This diff moves IOError exception handling into the
read loop and when an IOError occurs, it reboots the serial connection.

I have yet to see one of these in the wild during half an hour of testing, but
would rather push and revert than not push and forget to publish this patch.